Fix: Woah one year later and it's broken
This was because the material library used to have a wrapper over the tile element I was using (in order for the wave effect to occur on mouse click), and in the past I had to select the offsetParent as the 'tile' to bypass the 'wave'. Times have changed and that's no longer the case. This commit fixed that up and things are back to normal. Of course I could improve upon this, but I want to keep it as is to represent where I was in my coding journey back in early 2017.